Homework Урок 3. Порождающие шаблоны. Pattern Factory
