+++ web-server/src/main/java/ru/geekbrains/RequestParser.java
@@ -4,32 +4,11 @@ import ru.geekbrains.domain.HttpRequest;
 
 import java.util.Deque;
 
-public class RequestParser {
+public interface RequestParser {
 
   /**
    * Парсиниг строки запроса
    */
-  public HttpRequest parse(Deque<String> rawRequest){
-    HttpRequest httpRequest = new HttpRequest();
+  HttpRequest parse(Deque<String> rawRequest);
 
-    httpRequest.setBody(rawRequest.pollLast());
-    String[] method = rawRequest.pollFirst().split(" ", 3);
-    httpRequest.setMethod(method[0]);
-    httpRequest.setUrl(method[1]);
-
-        while (!rawRequest.isEmpty()) {
-            String line = rawRequest.pollFirst();
-            if (line.isBlank()) {
-                break;
-            }
-            String[] header = line.split(": ");
-            httpRequest.getHeaders().put(header[0], header[1]);
-        }
-        StringBuilder sb = new StringBuilder();
-        while (!rawRequest.isEmpty()) {
-            sb.append(rawRequest.pollFirst());
-        }
-        httpRequest.setBody(sb.toString());
-        return httpRequest;
-    }
 }
