+++ web-server/src/main/java/ru/geekbrains/RequestParserFactory.java
@@ -1,0 +1,11 @@
+package ru.geekbrains;
+
+/**
+ * Pattern Factory для сервиса RequestParser (разбор строки запроса от клиента)
+ */
+public final class RequestParserFactory {
+
+  public static RequestParser create(){
+    return new RequestParserImpl();
+  }
+}
