+++ web-server/src/main/java/ru/geekbrains/ResponseSerializerFactory.java
@@ -1,0 +1,10 @@
+package ru.geekbrains;
+
+/**
+ * Pattern Factory для сервиса ResponseSerializer (подготовка ответа клиенту)
+ */
+public final class ResponseSerializerFactory {
+  public static ResponseSerializer create(){
+    return new ResponseSerializerImpl();
+  }
+}
