+++ web-server/src/main/java/ru/geekbrains/ResponseSerializerImpl.java
@@ -1,0 +1,18 @@
+package ru.geekbrains;
+
+import ru.geekbrains.domain.HttpResponse;
+
+class ResponseSerializerImpl implements ResponseSerializer{
+
+    /**
+     * Объединение в ответ
+     */
+    public String serialize(HttpResponse response) {
+        StringBuilder sb = new StringBuilder();
+        sb.append("HTTP/1.1 " + response.getStatusCode() + " " + response.getStatusCodeName() + "\n");
+        response.getHeaders().forEach((header, value) -> sb.append(header + ": " + value + "\n"));
+        sb.append('\n');
+        sb.append(response.getBody());
+        return sb.toString();
+    }
+}
