+++ web-server/src/main/java/ru/geekbrains/service/FileServiceFactory.java
@@ -1,0 +1,11 @@
+package ru.geekbrains.service;
+
+/**
+ * Pattern Factory для сервиса FileService
+ */
+public final class FileServiceFactory {
+
+  public static FileService create(String rootDir) {
+    return new FileServiceImpl(rootDir);
+  }
+}
