+++ web-server/src/main/java/ru/geekbrains/service/SocketService.java
@@ -1,64 +1,22 @@
 package ru.geekbrains.service;
 
-import java.io.*;
-import java.net.Socket;
-import java.nio.charset.StandardCharsets;
+import java.io.IOException;
 import java.util.Deque;
-import java.util.LinkedList;
 
-public class SocketService implements Closeable {
+public interface SocketService {
 
-    private final Socket socket;
+  /**
+   * Чтение данных запроса от клиента
+   */
+  public Deque<String> readRequest();
 
-    public SocketService(Socket socket) {
-        this.socket = socket;
-    }
+  /**
+   * Отправка ответа клиенту
+   */
+  void writeResponse(String rawResponse);
 
-    /**
-     * Чтение данных запроса от клиента
-     */
-    public Deque<String> readRequest() {
-        try {
-            // открываем поток
-            BufferedReader input = new BufferedReader(
-                    new InputStreamReader(
-                            socket.getInputStream(), StandardCharsets.UTF_8));
-
-            // Ожидаем пока не придут данные из сокета
-            while (!input.ready());
-
-            Deque<String> response = new LinkedList<>();
-
-            // Ожидаем пока не придут данные из сокета
-            while (input.ready()) {
-                String line = input.readLine();
-                System.out.println("LOG: " + line);
-                response.add(line);
-            }
-            return response;
-        } catch (IOException ex) {
-            throw new IllegalStateException(ex);
-        }
-    }
-
-    /**
-     * Отправка ответа клиенту
-     */
-    public void writeResponse(String rawResponse) {
-        try {
-            PrintWriter output = new PrintWriter(socket.getOutputStream());
-            output.print(rawResponse);
-            output.flush();
-        } catch (IOException ex) {
-            throw new IllegalStateException(ex);
-        }
-    }
-
-    @Override
-    public void close() throws IOException {
-        if (!socket.isClosed()) {
-            socket.close();
-            System.out.println("LOG: Socket close!");
-        }
-    }
+  /**
+   * Закрытие соединения Socket
+   */
+  void close() throws IOException;
 }
