+++ web-server/src/main/java/ru/geekbrains/service/SocketServiceFactory.java
@@ -1,0 +1,12 @@
+package ru.geekbrains.service;
+
+import java.net.Socket;
+
+/**
+ * Pattern Factory для сервиса SocketService
+ */
+public final class SocketServiceFactory {
+  public static SocketService create(Socket socket) {
+    return new SocketServiceImpl(socket);
+  }
+}
