+++ web-server/src/test/resources/test.txt
@@ -1,0 +1,1 @@
+This file is test!
